+++ RELEASE_NOTES.md
@@ -1,5 +1,5 @@
 ### 1.1.10
-* added SQPNP solver kind
+* added SQPNP solver kind 
 
 ### 1.1.9
 * Added SURF, SIFT etc.